requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
psycopg2-binary==2.9.11
pycparser==2.23
PySocks==1.7.1
requests==2.32.5
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
undetected-chromedriver==3.5.5
urllib3==2.6.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
celery
redis
gunicorn
spacy
fastapi
uvicorn[standard]
pydantic
pyjwt
